+++ src/main/java/lotto/domain/winnerlotto/LottoWinner.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.winnerlotto;
 
 import java.util.Objects;
