+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
