+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lottoticket.LottoTicket;
+import lotto.domain.lottoticket.LottoTickets;
 import lotto.service.LottoTicketService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
