+++ src/test/java/lotto/domain/LottoWinnerBonusNumberTest.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.domain.winnerlotto.LottoWinnerTicket;
+import lotto.domain.number.LottoNumber;
+import lotto.domain.winnerlotto.LottoWinnerBonusNumber;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
