+++ src/test/java/lotto/domain/LottoWinnerTest.java
@@ -1,5 +1,9 @@
 package lotto.domain;
 
+import lotto.domain.winnerlotto.LottoWinnerTicket;
+import lotto.domain.number.LottoNumber;
+import lotto.domain.winnerlotto.LottoWinnerBonusNumber;
+import lotto.domain.winnerlotto.LottoWinner;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
